requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
biopython==1.86
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
dotenv==0.9.9
executing==2.2.1
fastuuid==0.14.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.2
idna==3.11
importlib_metadata==8.7.0
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
litellm==1.79.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
multidict==6.7.0
nest-asyncio==1.6.0
numpy==2.3.4
openai==2.7.1
packaging==25.0
parso==0.8.5
platformdirs==4.5.0
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.1.3
pure_eval==0.2.3
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.28.0
semlib==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tiktoken==0.12.0
tokenizers==0.22.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.14
yarl==1.22.0
zipp==3.23.0
